Move active_support tests to an integration suite
