+++ lib/hashie/extensions/deep_merge.rb
@@ -38,6 +38,8 @@ module Hashie
           hash[k] =
             if hash.key?(k) && hash[k].is_a?(::Hash) && v.is_a?(::Hash)
               _recursive_merge(hash[k], v, &block)
+            elsif v.is_a?(::Hash)
+              _recursive_merge({}, v, &block)
             elsif hash.key?(k) && block_given?
               yield(k, hash[k], v)
             else
