+++ spec/integration/active_support/Gemfile
@@ -1,0 +1,15 @@
+source 'http://rubygems.org'
+
+gem 'hashie', path: '../../..'
+require File.expand_path('../../../../lib/hashie/extensions/ruby_version', __FILE__)
+
+# rubocop:disable Bundler/DuplicatedGem
+if Hashie::Extensions::RubyVersion.new(RUBY_VERSION) >= Hashie::Extensions::RubyVersion.new('2.4.0')
+  gem 'activesupport', '~> 5.x', require: false
+else
+  gem 'activesupport', '~> 4.x', require: false
+end
+# rubocop:enable Bundler/DuplicatedGem
+
+gem 'rake'
+gem 'rspec', '~> 3.5.0'
